pyproject.toml
[build-system]
requires = ["setuptools"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "--cov=artrgb --cov-report html"

[tool.coverage.run]
omit=["*/tests/*"]

[project]
name = "pythonartrgb"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["Pillow", "numpy"]
dynamic = ["version"]
license = "MIT"
classifiers = ["Private :: Do Not Upload"]

# uv pip install -e '.[test|dev]'
[project.optional-dependencies]
test = ["pytest", "pytest-cov", "tox"]
dev = ["black", "isort"]


[tool.setuptools]
packages = ["ArtRgb"]
package-dir = { "" = "src" }
